+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.2.0
+## Added
+- Add ticket printing functionality in bitmap format.
+
 # v1.1.0
 ## Added
 - Add Refunds example using the public MP API